cleanup and clippy
