+++ contracts/core/manager/tests/add_ons.rs
@@ -1,8 +1,8 @@
 mod common;
 
-use common::{create_default_os, init_abstract_env, AResult, TEST_COIN};
 use abstract_boot::*;
 use abstract_os::PROXY;
+use common::{create_default_os, init_abstract_env, AResult, TEST_COIN};
 
 use boot_core::prelude::{instantiate_default_mock_env, ContractInstance};
 use cosmwasm_std::{Addr, Coin, CosmosMsg};
