+++ contracts/modules/perks/vesting/src/contract.rs
@@ -1,5 +1,3 @@
-
-
 use crate::queries::*;
 use abstract_sdk::os::{
     vesting::{
