--- contracts/native/module-factory/src/tests/README.md
@@ -1,12 +1,0 @@
-# Tests covered
-
-## Unit tests
-
-- Messages
-
-## Integration tests
-
-- Messages
----
-
-# Coverage
