--- contracts/native/module-factory/src/tests/common/mod.rs
@@ -1,5 +1,0 @@
-pub const PROXY_CONTRACT: &str = "proxy_addr";
-pub const ROOT_USER: &str = "root_user";
-pub const ANS_HOST_CONTRACT: &str = "ans_host_addr";
-pub const TEST_CREATOR: &str = "creator";
-pub const DEFAULT_VERSION: &str = "1.0.0";
