--- contracts/native/module-factory/src/tests/integration_tests/direct_calls.rs
@@ -1,48 +1,0 @@
-use super::init::init_abstract_env;
-use crate::tests::common;
-use abstract_boot::{OsFactoryQueryFns, VCQueryFns, *};
-use abstract_os::{module_factory, objects::module::ModuleInfo};
-use boot_core::prelude::{instantiate_default_mock_env, ContractInstance};
-use cosmwasm_std::Addr;
-use speculoos::prelude::*;
-
-type AResult = anyhow::Result<()>; // alias for Result<(), anyhow::Error>
-
-#[test]
-fn instantiate() -> AResult {
-    let sender = Addr::unchecked(common::ROOT_USER);
-    let (_state, chain) = instantiate_default_mock_env(&sender)?;
-    let (mut deployment, mut core) = init_abstract_env(&chain)?;
-    deployment.deploy(&mut core)?;
-
-    let factory = deployment.module_factory;
-    let factory_config = factory.config()?;
-    let expected = module_factory::ConfigResponse {
-        owner: sender.clone().into_string(),
-        ans_host_address: deployment.ans_host.address()?.into(),
-        version_control_address: deployment.version_control.address()?.into_string(),
-    };
-
-    assert_that!(&factory_config).is_equal_to(&expected);
-    Ok(())
-}
-
-#[test]
-fn caller_must_be_manager() -> AResult {
-    let _not_owner = Addr::unchecked("not_owner");
-    let sender = Addr::unchecked(common::ROOT_USER);
-    let (_, chain) = instantiate_default_mock_env(&sender)?;
-    let (mut deployment, mut core) = init_abstract_env(&chain)?;
-    deployment.deploy(&mut core)?;
-
-    let factory = &deployment.module_factory;
-    let test_module = ModuleInfo::from_id(
-        "publisher:test",
-        abstract_os::objects::module::ModuleVersion::Latest {},
-    )?;
-
-    let res = factory.install_module(test_module, None).unwrap_err();
-    assert_that(&res.root().to_string()).contains("Caller must be an OS manager");
-
-    Ok(())
-}
