--- contracts/native/module-factory/src/tests/integration_tests/mod.rs
@@ -1,2 +1,0 @@
-mod direct_calls;
-mod init;
