+++ contracts/native/module-factory/tests/common/mod.rs
@@ -1,5 +1,5 @@
 use abstract_boot::{
-    os_factory::OSFactory, AnsHost, Deployment, Manager, ModuleFactory, Proxy, VersionControl, OS,
+    AnsHost, Deployment, Manager, ModuleFactory, OSFactory, Proxy, VersionControl, OS,
 };
 
 use abstract_os::{ANS_HOST, MANAGER, MODULE_FACTORY, OS_FACTORY, PROXY, VERSION_CONTROL};
@@ -7,12 +7,7 @@ use boot_core::{prelude::ContractInstance, Mock};
 
 use cw_multi_test::ContractWrapper;
 
-pub const PROXY_CONTRACT: &str = "proxy_addr";
 pub const ROOT_USER: &str = "root_user";
-pub const ANS_HOST_CONTRACT: &str = "ans_host_addr";
-pub const TEST_CREATOR: &str = "creator";
-pub const DEFAULT_VERSION: &str = "1.0.0";
-
 
 pub fn init_test_env<'a>(chain: &'a Mock) -> anyhow::Result<(Deployment<'a, Mock>, OS<Mock>)> {
     let mut ans_host = AnsHost::new(ANS_HOST, chain);
