+++ contracts/testing/src/tests/common/mod.rs
@@ -1,10 +1,6 @@
-pub const TREASURY_CONTRACT: &str = "proxy_contract_address";
-pub const ROOT_USER: &str = "trader_contract_address";
-pub const ANS_HOST_CONTRACT: &str = "ans_host_contract_address";
 pub const TEST_CREATOR: &str = "creator";
 pub const RANDOM_USER: &str = "random_user";
 pub const DEFAULT_VERSION: &str = "0.1.0";
-pub const CHAIN_ID: &str = "cosmos-testnet-14002";
 pub const OS_NAME: &str = "default_name";
 pub const SUBSCRIPTION_COST: &str = "0.1";
 pub const DEFAULT_PAY: &str = "0.01";
