+++ packages/abstract-sdk/src/apis/verify.rs
@@ -53,7 +53,9 @@ impl<'a, T: Verification> OsRegister<'a, T> {
 
     /// Verify if the provided proxy address is indeed a user.
     pub fn assert_proxy(&self, maybe_proxy: &Addr) -> StdResult<Core> {
-        let os_id = OS_ID.query(&self.deps.querier, maybe_proxy.clone()).map_err(|_| StdError::generic_err("Caller must be an OS proxy."))?;
+        let os_id = OS_ID
+            .query(&self.deps.querier, maybe_proxy.clone())
+            .map_err(|_| StdError::generic_err("Caller must be an OS proxy."))?;
         let maybe_os =
             OS_ADDRESSES.query(&self.deps.querier, self.base.registry(self.deps)?, os_id)?;
         match maybe_os {
